Complete remake with quantities defined by base quantities
A Quantity should consist of a "quotient" of two lists
of Quantities, e.g.:
speed = (distance) / (duration)
force = (mass, distance, distance) / (duration)

Adding/subtracting is only possible between two equal quantities.

Multiplication/Division is possible between any quantities. The
resulting numerator and denominator is the union (inverse union for
division) of the two operands' num. and denom.. The resulting num. and
denom. are then simplified (each pair of the same base quantity in each
is removed).

It's not needed to define quantities or units for "inbetween" results,
these can be left "anonymous". The only thing that defines a quantity
are the base quantities in the quotient.

Made wrong usage as difficult as possible by adding concepts and
constraints on the types used as base quantities and the quantities used
in operators.

Added tests for operations that were previously impossible.

Improved Readme and tests.

Moved to conan2.